api_yamdb/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via
    #   -r requirements.in
    #   django
atomicwrites==1.4.1
    # via -r requirements.in
attrs==22.2.0
    # via
    #   -r requirements.in
    #   pytest
black==23.1.0
    # via -r requirements.in
build==0.10.0
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2022.12.7
    # via
    #   -r requirements.in
    #   requests
cfgv==3.3.1
    # via
    #   -r requirements.in
    #   pre-commit
charset-normalizer==2.0.12
    # via
    #   -r requirements.in
    #   requests
click==8.1.3
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
colorama==0.4.6
    # via -r requirements.in
distlib==0.3.6
    # via
    #   -r requirements.in
    #   virtualenv
django==3.2
    # via
    #   -r requirements.in
    #   django-filter
    #   django-model-utils
    #   djangorestframework
    #   djangorestframework-simplejwt
django-filter==22.1
    # via -r requirements.in
django-model-utils==4.3.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
filelock==3.10.0
    # via
    #   -r requirements.in
    #   virtualenv
flake8==6.0.0
    # via -r requirements.in
identify==2.5.21
    # via
    #   -r requirements.in
    #   pre-commit
idna==3.4
    # via
    #   -r requirements.in
    #   requests
iniconfig==2.0.0
    # via
    #   -r requirements.in
    #   pytest
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   flake8
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
nodeenv==1.7.0
    # via
    #   -r requirements.in
    #   pre-commit
packaging==23.0
    # via
    #   -r requirements.in
    #   black
    #   build
    #   pytest
pathspec==0.11.1
    # via
    #   -r requirements.in
    #   black
pip-tools==6.12.3
    # via -r requirements.in
platformdirs==3.1.1
    # via
    #   -r requirements.in
    #   black
    #   virtualenv
pluggy==0.13.1
    # via
    #   -r requirements.in
    #   pytest
pre-commit==3.2.0
    # via -r requirements.in
py==1.11.0
    # via
    #   -r requirements.in
    #   pytest
pycodestyle==2.10.0
    # via
    #   -r requirements.in
    #   flake8
pyflakes==3.0.1
    # via
    #   -r requirements.in
    #   flake8
pyjwt==2.1.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
pytest==6.2.4
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-pythonpath
pytest-django==4.4.0
    # via -r requirements.in
pytest-pythonpath==0.7.3
    # via -r requirements.in
pytz==2022.7.1
    # via
    #   -r requirements.in
    #   django
pyyaml==6.0
    # via
    #   -r requirements.in
    #   pre-commit
requests==2.26.0
    # via -r requirements.in
sqlparse==0.4.3
    # via
    #   -r requirements.in
    #   django
toml==0.10.2
    # via
    #   -r requirements.in
    #   pytest
tomli==2.0.1
    # via
    #   -r requirements.in
    #   black
    #   build
    #   pyproject-hooks
typing-extensions==4.5.0
    # via
    #   -r requirements.in
    #   black
urllib3==1.26.15
    # via
    #   -r requirements.in
    #   requests
virtualenv==20.21.0
    # via
    #   -r requirements.in
    #   pre-commit
wheel==0.40.0
    # via pip-tools
gunicorn
psycopg2-binary
# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
